create core module and expose only relevant lib interface
